Break the dao trait into a separate file
