+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/versioning/IngestVersionManagerDao.scala
@@ -1,0 +1,15 @@
+package uk.ac.wellcome.platform.archive.common.versioning
+
+import uk.ac.wellcome.platform.archive.common.IngestID
+import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
+
+import scala.util.Try
+
+trait IngestVersionManagerDao {
+  def lookupExistingVersion(ingestID: IngestID): Try[Option[VersionRecord]]
+
+  def lookupLatestVersionFor(
+    externalIdentifier: ExternalIdentifier): Try[Option[VersionRecord]]
+
+  def storeNewVersion(record: VersionRecord): Try[Unit]
+}
